requirements.txt
Flask
requests
gunicorn
Werkzeug
numpy
tensorflow==2.15.0
keras==2.15.0
Pillow